fix bug mean is 0

+++ hw2/hw2/PercolationStats.java
@@ -40,7 +40,7 @@ public class PercolationStats {
                     percolation.open(row, col);
                 }
             }
-            double threshold = percolation.numberOfOpenSites() / (N * N);
+            double threshold = (double) percolation.numberOfOpenSites() / (N * N);
             thresholds[t] = threshold;
         }
     }
